Hack around umac_ctx name-punning
OpenSSH builds the file umac.c twice, once as is, and once using
the preprocessor to rename everything to umac128_*.

Unfortunately, it does not rename the struct umac_ctx, which ends
up being used by both .o files to talk about completely different
structs (the preprocessor macros change its size).

This only matters if you care about DWARF/CTF type information,
which we do, so we add an extra -D to rename the struct in the
second build of umac.c.