Per-session xauthfile
This patch is to fix a X11 connection failure when a user's home directory
is read-only.

We have contributed back this fix to the OpenSSH upstream community. For
more information, see https://bugzilla.mindrot.org/show_bug.cgi?id=2440
In the future, if this fix is accepted by the upsteam in a later release, we
will remove this patch when we upgrade to that release.